requirements/dev.txt
-r prod.txt

black==25.1.0
django-debug-toolbar==5.2.0
flake8==7.1.1
flake8-absolute-import==1.0.0.2
flake8-broken-line==1.0.0
flake8-bugbear==24.12.12
flake8-builtins==2.5.0
flake8-clean-block==0.1.2
flake8-commas==4.0.0
flake8-comments==0.1.2
flake8-comprehensions==3.16.0
flake8-continuation==1.0.5
flake8-dunder-all==0.4.1
flake8-eradicate==1.5.0
flake8-expression-complexity==0.0.11
flake8-fixme==1.1.1
flake8-forbidden-func==0.1.0
flake8-implicit-str-concat==0.5.0
flake8-import-guard==1.0.0
flake8-import-order==0.18.2
flake8-mutable==1.2.0
flake8-print==5.0.0
Flake8-pyproject==1.2.3
flake8-quotes==3.4.0
flake8-return==1.2.0
flake8-type-ignore==0.1.0.post2
flake8-use-pathlib==0.3.0
pep8-naming==0.14.1
